PSG-2683: rename jenkins dir with integration_tests
