+++ integration_tests/files/sars_cov_2/expected_results/pipeline_ci_illumina/results.csv
