+++ integration_tests/files/synthetic/expected_results/pipeline_ci_illumina/results.csv
