--- jenkins/files/s_aureus/README
@@ -1,1 +1,0 @@
-Add jenkinsfiles and expected results to this folder
