--- jenkins/files/sars_cov_2/pipeline_ci_tests/Jenkinsfile
@@ -1,87 +1,0 @@
-/*
- * processing time: few minutes
- */
-pipeline {
-    agent none
-    environment {
-        PSGA_PIPELINE = "psga-pipeline-sars-cov-2"
-    }
-    stages {
-        stage('Run PSGA pipeline CI tests') {
-            parallel {
-                stage('ILLUMINA/unknown (mix primers)') {
-                    steps {
-                        script {
-                            build (
-                                job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                parameters: [
-                                        booleanParam(name: 'VALIDATION_TEST', value: true),
-                                        string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/pipeline_ci_tests/illumina/metadata.csv"),
-                                        string(name: 'OUTPUT_PATH', value: "/data/output/pipeline_ci_illumina"),
-                                        string(name: 'NXF_WORK', value: "/data/work/pipeline_ci_illumina"),
-                                        string(name: 'ANALYSIS_RUN', value: "pipeline_ci_illumina"),
-                                        string(name: 'SEQUENCING_TECHNOLOGY', value: "illumina"),
-                                        string(name: 'KIT', value: "unknown"),
-                                ]
-                            )
-                        }
-                    }
-                }
-                stage('ONT/unknown (mix primers)') {
-                    steps {
-                        script {
-                            build (
-                                job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                parameters: [
-                                        booleanParam(name: 'VALIDATION_TEST', value: true),
-                                        string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/pipeline_ci_tests/ont/metadata.csv"),
-                                        string(name: 'OUTPUT_PATH', value: "/data/output/pipeline_ci_ont"),
-                                        string(name: 'NXF_WORK', value: "/data/work/pipeline_ci_ont"),
-                                        string(name: 'ANALYSIS_RUN', value: "pipeline_ci_ont"),
-                                        string(name: 'SEQUENCING_TECHNOLOGY', value: "ont"),
-                                        string(name: 'KIT', value: "unknown"),
-                                ]
-                            )
-                        }
-                    }
-                }
-                stage('FASTA') {
-                    steps {
-                        script {
-                            build (
-                                job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                parameters: [
-                                        booleanParam(name: 'VALIDATION_TEST', value: true),
-                                        string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/pipeline_ci_tests/unknown/metadata.csv"),
-                                        string(name: 'OUTPUT_PATH', value: "/data/output/pipeline_ci_unknown"),
-                                        string(name: 'NXF_WORK', value: "/data/work/pipeline_ci_unknown"),
-                                        string(name: 'ANALYSIS_RUN', value: "pipeline_ci_unknown"),
-                                        string(name: 'SEQUENCING_TECHNOLOGY', value: "unknown"),
-                                        string(name: 'KIT', value: "none"),
-                                ]
-                            )
-                        }
-                    }
-                }
-                stage('FAILING_ILLUMINA_SAMPLES') {
-                    steps {
-                        script {
-                            build (
-                                job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                parameters: [
-                                        booleanParam(name: 'VALIDATION_TEST', value: false),
-                                        string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/from_them/gold_standard/illumina/v4/2022-08-18/pipeline_metadata.csv"),
-                                        string(name: 'OUTPUT_PATH', value: "/data/output/empty_sample_and_sample_failing_ncov_qc"),
-                                        string(name: 'NXF_WORK', value: "/data/work/empty_sample_and_sample_failing_ncov_qc"),
-                                        string(name: 'ANALYSIS_RUN', value: "empty_sample_and_sample_failing_ncov_qc"),
-                                        string(name: 'SEQUENCING_TECHNOLOGY', value: "illumina"),
-                                        string(name: 'KIT', value: "ARTIC_V4-1"),
-                                ]
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
