--- jenkins/files/sars_cov_2/small_tests/Jenkinsfile
@@ -1,123 +1,0 @@
-/*
- * processing time: few minutes
- */
-pipeline {
-    agent none
-    environment {
-        PSGA_PIPELINE = "psga-pipeline-sars-cov-2"
-    }
-    stages {
-        stage('Run PSGA pipeline: small tests') {
-            parallel {
-                stage('ILLUMINA/unknown (ARTIC_V4-1)') {
-                    steps {
-                        script {
-                            build (
-                                job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                parameters: [
-                                        booleanParam(name: 'VALIDATION_TEST', value: true),
-                                        string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/small_tests/illumina/metadata.csv"),
-                                        string(name: 'OUTPUT_PATH', value: "/data/output/illumina_artic_v4-1"),
-                                        string(name: 'NXF_WORK', value: "/data/work/illumina_artic_v4-1"),
-                                        string(name: 'ANALYSIS_RUN', value: "illumina_artic_v4-1"),
-                                        string(name: 'SEQUENCING_TECHNOLOGY', value: "illumina"),
-                                        string(name: 'KIT', value: "unknown"),
-                                ]
-                            )
-                        }
-                    }
-                }
-                stage('ONT/none (set to none for the time being. This is NOT ARTIC for sure)') {
-                    steps {
-                        script {
-                            build (
-                                job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                parameters: [
-                                        booleanParam(name: 'VALIDATION_TEST', value: true),
-                                        string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/small_tests/ont/metadata.csv"),
-                                        string(name: 'OUTPUT_PATH', value: "/data/output/ont_none"),
-                                        string(name: 'NXF_WORK', value: "/data/work/ont_none"),
-                                        string(name: 'ANALYSIS_RUN', value: "ont_none"),
-                                        string(name: 'SEQUENCING_TECHNOLOGY', value: "ont"),
-                                        string(name: 'KIT', value: "none"),
-                                ]
-                            )
-                        }
-                    }
-                }
-                stage('ONT/unknown (ARTIC_V4-1)') {
-                    steps {
-                        script {
-                            build (
-                                job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                parameters: [
-                                        booleanParam(name: 'VALIDATION_TEST', value: true),
-                                        string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/requirements/CL0015b/ont_v4_input/metadata.csv"),
-                                        string(name: 'OUTPUT_PATH', value: "/data/output/ont_artic_v4-1"),
-                                        string(name: 'NXF_WORK', value: "/data/work/ont_artic_v4-1"),
-                                        string(name: 'ANALYSIS_RUN', value: "ont_artic_v4-1"),
-                                        string(name: 'SEQUENCING_TECHNOLOGY', value: "ont"),
-                                        string(name: 'KIT', value: "unknown"),
-                                ]
-                            )
-                        }
-                    }
-                }
-                stage('ONT/MIDNIGHT_V2') {
-                    steps {
-                        script {
-                            build (
-                                job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                parameters: [
-                                        booleanParam(name: 'VALIDATION_TEST', value: true),
-                                        string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/small_tests/ont_midnight_v2/metadata.csv"),
-                                        string(name: 'OUTPUT_PATH', value: "/data/output/ont_midnight_v2"),
-                                        string(name: 'NXF_WORK', value: "/data/work/ont_midnight_v2"),
-                                        string(name: 'ANALYSIS_RUN', value: "ont_midnight_v2"),
-                                        string(name: 'SEQUENCING_TECHNOLOGY', value: "ont"),
-                                        string(name: 'KIT', value: "Midnight-ONT_V2"),
-                                ]
-                            )
-                        }
-                    }
-                }
-                stage('FASTA') {
-                    steps {
-                        script {
-                            build (
-                                job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                parameters: [
-                                        booleanParam(name: 'VALIDATION_TEST', value: true),
-                                        string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/small_tests/unknown/metadata.csv"),
-                                        string(name: 'OUTPUT_PATH', value: "/data/output/unknown_short"),
-                                        string(name: 'NXF_WORK', value: "/data/work/unknown_short"),
-                                        string(name: 'ANALYSIS_RUN', value: "unknown_short"),
-                                        string(name: 'SEQUENCING_TECHNOLOGY', value: "unknown"),
-                                        string(name: 'KIT', value: "none"),
-                                ]
-                            )
-                        }
-                    }
-                }
-                stage('FAILING_ILLUMINA_SAMPLES') {
-                    steps {
-                        script {
-                            build (
-                                job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                parameters: [
-                                        booleanParam(name: 'VALIDATION_TEST', value: false),
-                                        string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/from_them/gold_standard/illumina/v4/2022-08-18/pipeline_metadata.csv"),
-                                        string(name: 'OUTPUT_PATH', value: "/data/output/empty_sample_and_sample_failing_ncov_qc"),
-                                        string(name: 'NXF_WORK', value: "/data/work/empty_sample_and_sample_failing_ncov_qc"),
-                                        string(name: 'ANALYSIS_RUN', value: "empty_sample_and_sample_failing_ncov_qc"),
-                                        string(name: 'SEQUENCING_TECHNOLOGY', value: "illumina"),
-                                        string(name: 'KIT', value: "ARTIC_V4-1"),
-                                ]
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
