--- jenkins/files/sars_cov_2/validation-ci/Jenkinsfile
@@ -1,88 +1,0 @@
-/*
- * samples per test: 30 (10 alpha, 10 delta, 10 omicron. ont_artic_fasta contains a mix of variants)
- * processing time: ~1.5 hours
- */
-
-/*
-Constructs the cron schedule for this build in Jenkins but only if the branch is dev. This avoids applying
-this cron schedule to all branches in this multi-branch pipeline.
-*/
-def getCronSchedule() {
-    return env.BRANCH_NAME == "dev" ? "H 0 * * 1-5" : ""
-}
-
-pipeline {
-    triggers {
-        cron(getCronSchedule())
-    }
-    agent none
-    environment {
-        PSGA_PIPELINE = "psga-pipeline-sars-cov-2"
-    }
-    stages {
-        stage('Run PSGA pipeline: validation-ci') {
-            stages {
-                stage('illumina') {
-                    steps {
-                        script {
-                            catchError(buildResult: 'SUCCESS', stageResult: 'FAILURE') {
-                                build (
-                                    job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                    parameters: [
-                                            booleanParam(name: 'VALIDATION_TEST', value: true),
-                                            string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/validation_ci/illumina/metadata.csv"),
-                                            string(name: 'OUTPUT_PATH', value: "/data/output/illumina"),
-                                            string(name: 'NXF_WORK', value: "/data/work/illumina"),
-                                            string(name: 'ANALYSIS_RUN', value: "illumina"),
-                                            string(name: 'SEQUENCING_TECHNOLOGY', value: "illumina"),
-                                            string(name: 'KIT', value: "ARTIC_V3"),
-                                    ]
-                                )
-                            }
-                        }
-                    }
-                }
-                stage('ont') {
-                    steps {
-                        script {
-                            catchError(buildResult: 'SUCCESS', stageResult: 'FAILURE') {
-                                build (
-                                    job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                    parameters: [
-                                            booleanParam(name: 'VALIDATION_TEST', value: true),
-                                            string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/validation_ci/ont/metadata.csv"),
-                                            string(name: 'OUTPUT_PATH', value: "/data/output/ont"),
-                                            string(name: 'NXF_WORK', value: "/data/work/ont"),
-                                            string(name: 'ANALYSIS_RUN', value: "ont"),
-                                            string(name: 'SEQUENCING_TECHNOLOGY', value: "ont"),
-                                            string(name: 'KIT', value: "ARTIC_V3"),
-                                    ]
-                                )
-                            }
-                        }
-                    }
-                }
-                stage('fasta') {
-                    steps {
-                        script {
-                            catchError(buildResult: 'SUCCESS', stageResult: 'FAILURE') {
-                                build (
-                                    job: "${PSGA_PIPELINE}/${env.BRANCH_NAME}",
-                                    parameters: [
-                                            booleanParam(name: 'VALIDATION_TEST', value: true),
-                                            string(name: 'METADATA', value: "s3://synthetic-data-dev/UKHSA/validation_ci/unknown/metadata.csv"),
-                                            string(name: 'OUTPUT_PATH', value: "/data/output/unknown"),
-                                            string(name: 'NXF_WORK', value: "/data/work/unknown"),
-                                            string(name: 'ANALYSIS_RUN', value: "unknown"),
-                                            string(name: 'SEQUENCING_TECHNOLOGY', value: "unknown"),
-                                            string(name: 'KIT', value: "none"),
-                                    ]
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
