+++ tests/integration_tests/test_loading.py
@@ -5,8 +5,8 @@ import pytest
 import pandas as pd
 
 from scripts.util.metadata import SAMPLE_ID
-from jenkins.loading import get_file_paths, load_data_from_csv
-from tests.jenkins.util import create_paths
+from integration_tests.loading import get_file_paths, load_data_from_csv
+from tests.integration_tests.util import create_paths
 
 
 @pytest.mark.parametrize("pathogen", ["sars_cov_2", "synthetic"])
